src/agent/requirements.txt
# Requirements for Weather-Based Clothing Advisor Agent
# Compatible with both Container Apps and Foundry Hosted deployments

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# HTTP client
requests>=2.28.0
httpx>=0.24.0

# Azure dependencies
azure-identity>=1.12.0
azure-monitor-opentelemetry>=1.0.0

# OpenTelemetry for observability
opentelemetry-api>=1.15.0
opentelemetry-sdk>=1.15.0
opentelemetry-instrumentation-fastapi>=0.36b0

# Microsoft Agent Framework
# Note: Install from GitHub until published to PyPI
# pip install git+https://github.com/microsoft/agent-framework.git
agent-framework-core>=0.1.0
agent-framework-azure-ai>=0.1.0

# For Foundry Hosted deployment
azure-ai-agentserver-agentframework>=0.1.0

# Data validation
pydantic>=2.0.0

# Async utilities
anyio>=3.6.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.0.270
